requirements.txt
fastapi
pydantic>=2.0
pydantic-ai
google-generativeai
lancedb
numpy
pandas
requests
httpx
python-dotenv
rich
azure-functions
uvicorn
pypdf
streamlit
ipykernel
streamlit
requests



